# api/openapi.yaml
openapi: 3.1.0
info:
  title: yamo
  version: "1.0.0"
  summary: Catalogue, search and edit the metadata of a large music library.
  description: |
    yamo keeps a catalogue of a music library and edits the tags in the files
    themselves. This API is the only way anything reaches those files: the
    terminal interface and the command line are both clients of it.

    ## Working with a large library

    The library is expected to run to six figures, and the work is gradual —
    find a few wrong things, fix them, carry on. Three parts of the design
    follow from that.

    **Everything is paged.** `GET /tracks` returns a window and a total. There
    is no endpoint that returns the whole library.

    ## Paging

    Paging is offset-based. `limit` and `offset` go in the query string; the
    response is an envelope of `items`, `total`, `limit` and `offset`.

    ```
    GET /v1/tracks?q=artist:elvis&sort=album,track&limit=50&offset=100
    → { "items": [ … 50 tracks … ], "total": 2400, "limit": 50, "offset": 100 }
    ```

    - `total` is the number of matches **before** paging, so it is what a
      result count or a scrollbar should be sized from.
    - `limit` and `offset` are echoed back as applied. A `limit` above the
      maximum is capped rather than rejected, so read the response rather than
      assuming the request was honoured.
    - You have reached the end when `offset + len(items) >= total`, or when
...
      as offset 0.

...

...
    not a page of a larger set.

...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /tracks:
    get:
      tags: [tracks]
      summary: List and search tracks
      operationId: listTracks
      parameters:
        - $ref: '#/components/parameters/Query'
        - name: sort
          in: query
          description: |
            Comma-separated field names, each optionally prefixed with `-` for
            descending. Unrecognised names are ignored rather than rejected, so
            a half-typed sort degrades instead of failing.

            `score` sorts by how well each track matched, best first — `-score`
            inverts it. A fuzzy query with no `sort` at all is already ordered
            that way; passing one replaces that ordering rather than adding to
            it.
          schema: { type: string, examples: ["artist,album,track", "-year", "score"] }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of tracks.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackPage' }
  /tracks/{id}:
    parameters:
      - $ref: '#/components/parameters/TrackId'
    get:
      tags: [tracks]
      summary: Fetch one track
      operationId: getTrack
      responses:
        '200':
          description: The track.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Track' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [tracks]
      summary: Edit one track
      description: |
        Applies a sparse set of changes and writes them to the file before
        returning. Fields absent from the body are left alone; fields present
        with a null value are cleared.

        Setting a field to the value it already has does not rewrite the file,
        so a repeated request is free rather than churning the library.
      operationId: patchTrack
      parameters:
        - name: If-Match
          in: header
          description: |
            The `version` the client last saw. When present it must still match,
            or the request fails with `409` because the file changed underneath.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Changes' }
            examples:
              correctArtist:
                summary: Fix an artist and clear a stray comment
                value: { artist: "Elvis Presley", comment: null }
      responses:
        '200':
          description: The updated track.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Track' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unwritable' }
    delete:
      tags: [tracks]
      summary: Delete the file
      description: |
        Removes the file from disk and drops it from the catalogue. This is a
        real deletion rather than a move to a holding area: the strip backup
        exists because tags removed from a file cannot be recovered from
        anywhere else, whereas whole files are what a NAS's own snapshots and
        recycle bin are for.

        A file already gone from disk is not an error. The catalogue is a
        snapshot and can name files removed outside this program, and clearing
        the entry is what was asked for either way.
      operationId: deleteTrack
      parameters:
        - name: If-Match
          in: header
          description: |
            The `version` the client last saw. When present it must still
            match, or the request fails with `409` because the file changed
            underneath. Worth sending here more than anywhere else: nothing
            inside this program can undo the request.
          schema: { type: string }
      responses:
        '204': { description: "The file was deleted, or was already gone." }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tracks/{id}/rename:
    parameters:
      - $ref: '#/components/parameters/TrackId'
    post:
      tags: [tracks]
      summary: Rename or move the file
      description: |
        Moves the file and follows it in the catalogue.

        `path` may be a bare name, which renames the file where it is; a
        relative path, which resolves against the track's own directory; or an
        absolute path. Missing parent directories are created, so moving a
        stray track into the album folder it belongs in is one request.

        Three destinations are refused. One outside every library root, so
        that this is not a way to scatter files across the machine the server
        runs on. One that already exists — `409` with code `exists`, which
        means "choose another name" rather than "read it again and retry". And
        one that changes the extension, because the catalogue records the
        format and a file renamed across containers would claim to be
        something it is not.

        **A track's id is derived from its path, so a rename changes it.** The
        response is the track under its new identity and `Location` points at
        it; the old id stops resolving. The `tracks.changed` event names both,
        so a client holding the old one knows to drop it.
      operationId: renameTrack
      parameters:
        - name: If-Match
          in: header
          description: |
            The `version` the client last saw. When present it must still
            match, or the request fails with `409`.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RenameRequest' }
            examples:
              renameInPlace:
                summary: Correct a filename
                value: { path: "03 Blue Suede Shoes.mp3" }
              moveIntoAlbum:
                summary: Move a stray track into its album folder
                value: { path: "../Sun Sessions/03 Blue Suede Shoes.mp3" }
      responses:
        '200':
          description: The track, under its new id.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Location:
              description: Where the track now answers, which is not where it was asked for.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Track' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/RenameConflict' }
  /tracks/{id}/artwork:
    parameters:
      - $ref: '#/components/parameters/TrackId'
    get:
      tags: [artwork]
      summary: Fetch a track's cover
      operationId: getArtwork
      responses:
        '200':
          description: The image.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
        '404':
          description: The track has no embedded artwork.
    put:
      tags: [artwork]
      summary: Replace a track's cover
      operationId: putArtwork
      requestBody:
        required: true
        description: The raw image. The format is taken from the content, not the header.
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
          application/octet-stream: { schema: { type: string, format: binary } }
      responses:
        '200':
          description: The updated track.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Track' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unwritable' }
    delete:
      tags: [artwork]
      summary: Remove a track's cover
      operationId: deleteArtwork
      responses:
        '204': { description: "The artwork was removed, or there was none." }
        '404': { $ref: '#/components/responses/NotFound' }
  /albums:
    get:
      tags: [browse]
      summary: List albums
      description: |
        Returns the albums **containing tracks that match the query** — the
        query filters tracks, and the matching tracks are then grouped.

        That is deliberate, and worth reading twice: `?q=cat` returns an album
        whose composer is "Jamie Catto", because a bare term substring-matches
        every text field of every track. It is how you find the album with that
        one song on it, but the album's own name need not contain the term. To
        search album titles alone, scope the term: `?q=album:cat`.

        `tracks` counts only the matching tracks, not the whole album. Each
        album carries a `query` that reselects exactly it, so a client can go
        from a search result to the full record in one more call.

        This is the natural entry point for a browsing interface, where the
        unit of work is usually a record rather than a track.
      operationId: listAlbums
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of albums.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlbumPage' }
  /values/{field}:
    get:
      tags: [browse]
      summary: Distinct values of a field
      description: |
        Powers autocomplete and faceted browsing. Values are returned most-used
        first, with the count of tracks carrying each. Prefix matches are
        preferred; when there are none, substring matches are returned so a
        partly-remembered album still finds itself.
      operationId: listValues
      parameters:
        - name: field
          in: path
          required: true
          schema: { $ref: '#/components/schemas/FieldName' }
        - name: prefix
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Matching values.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ValueCount' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /tracks/batch:
    post:
      tags: [batch]
      summary: Apply changes to many tracks
      description: |
        The operation this API exists for. Applies one set of changes to every
        selected track, where the selection is usually a query rather than a
        list of identifiers.
      operationId: batchEditTracks
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchEditRequest' }
            examples:
              setArtist:
                summary: Set the artist across a whole search
                value:
                  selector: { query: "artist:\"elvis presly\"", expectCount: 42 }
                  set: { artist: "Elvis Presley" }
      responses:
        '202': { $ref: '#/components/responses/JobAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/CountMismatch' }
  /tracks/split:
    post:
      tags: [batch]
      summary: Split each title into the fields it contains
      description: |
        For a compilation whose artist tag says Various Artists and whose titles
        therefore carry the performer: "Michael Jackson - Beat It". The template
        describes the shape the title is in, naming fields with a leading `$`,
        and every capture but the last stops at the first separator that matches
        in full — so `$artist - $title` gets "Jay-Z - 99 Problems" right, and
        gives "Faithless - Insomnia - Monster Mix" the whole remainder as its
        title.

        Separators are matched literally. Values are trimmed, an empty one is
        not written, and a title the template does not fit is counted in
        `unmatched` and left alone rather than guessed at.

        Run it with `dryRun` first: the result carries worked examples and the
        count of titles that did not fit, which together are what say whether
        the template is right before it is applied to everything selected.
      operationId: splitTracks
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SplitRequest' }
            examples:
              artistDashTitle:
                summary: Pull the performer out of a compilation's titles
                value:
                  selector: { query: "album:\"now that's what i call music\"" }
                  template: "$artist - $title"
                  dryRun: true
      responses:
        '202': { $ref: '#/components/responses/JobAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/CountMismatch' }
  /artwork/batch:
    post:
      tags: [artwork, batch]
      summary: Set or clear artwork across many tracks
      description: |
        `folder` is the one worth knowing about: it embeds the `cover.jpg` or
        `folder.jpg` sitting beside each track, which is how a downloaded
        library normally stores art and the usual reason none of it appears on
        a phone. Each directory is read once, not once per track.

        Embedding artwork rewrites the file, since a cover is far larger than
        the padding any format reserves. Tracks whose art already matches are
        skipped, so a repeated run costs nothing.
      operationId: batchArtwork
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchArtworkRequest' }
      responses:
        '202': { $ref: '#/components/responses/JobAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/CountMismatch' }
  /clipboard/artwork:
    get:
      tags: [artwork]
      summary: Read the artwork clipboard
      operationId: getClipboard
      responses:
        '200':
          description: The held image.
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
        '404': { description: Nothing has been copied. }
    put:
      tags: [artwork]
      summary: Copy an image to the clipboard
      description: |
        The clipboard lives on the server, so a cover copied in the terminal can
        be pasted from a phone and the other way round.
      operationId: putClipboard
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
          application/octet-stream: { schema: { type: string, format: binary } }
      responses:
        '200':
          description: What is now held.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Picture' }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      tags: [artwork]
      summary: Empty the clipboard
      operationId: deleteClipboard
      responses:
        '204': { description: Emptied. }
  /clipboard/artwork/from-track/{id}:
    put:
      tags: [artwork]
      parameters:
        - $ref: '#/components/parameters/TrackId'
      summary: Copy a track's cover to the clipboard
      operationId: copyArtworkFromTrack
      responses:
        '200':
          description: What is now held.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Picture' }
        '404': { $ref: '#/components/responses/NotFound' }
  /clipboard/artwork/from-url:
    put:
      tags: [artwork, discogs]
      summary: Copy a Discogs cover to the clipboard
      description: |
        Downloads a cover from Discogs and holds it, from where the ordinary
        paste (`source: clipboard` on `/artwork/batch`) applies it to one track
        or to a whole album.

        The download happens here rather than in the browser because Discogs'
        image host sends no `Access-Control-Allow-Origin`: a page can display
        one of those URLs in an `<img>` but cannot read its bytes.

        The URL must be on a Discogs image host. Anything else is refused, so
        this cannot be used to make the server fetch an arbitrary address.
      operationId: copyArtworkFromURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  description: An image URL from a Discogs lookup.
                  examples: ["https://i.discogs.com/xxxx/rs:fit/g:sm/q:90/h:600/w:600/abc.jpeg"]
      responses:
        '200':
          description: What is now held.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Picture' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/LookupUnavailable' }
  /tracks/{id}/audio:
    get:
      tags: [tracks]
      summary: The audio itself
      description: |
        Streams the file, so a player can confirm that the song in front of you
        is the one the tags describe.

        Range requests are answered, which is what lets a player seek without
        pulling the whole file first. The media type comes from the format the
        catalogue recorded rather than from sniffing the bytes.

        A browser cannot point an `<audio>` element straight at this when the
        page and the server are different origins: the element sends no
        `Authorization` header and a token is required cross-origin. Fetch it
        and play the blob instead, as the bundled web app does.
      operationId: getAudio
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: The whole file.
          content:
            audio/*:
              schema: { type: string, format: binary }
        '206':
          description: The requested byte range.
          content:
            audio/*:
              schema: { type: string, format: binary }
        '404':
          $ref: '#/components/responses/NotFound'
  /discogs/search:
    get:
      tags: [discogs]
      summary: Find album covers on Discogs
      description: |
        Searches Discogs for **master** releases — a master is the album, with
        one entry covering every pressing and reissue, so a release search
        would return the same sleeve twenty times over.

        Costs more than one request. An unauthenticated Discogs search returns
        empty `thumb` and `cover_image` fields, so each candidate has to be
        fetched by id to discover its cover: a search is one request plus one
        per candidate, against a budget of 25 a minute. Candidates are capped
        for that reason, results are cached, and `rateRemaining` reports what
        is left so a client can pace itself. Configure a token on the server to
        raise the budget to 60 and get covers from the search itself.

        Hits that turn out to have no image at all are omitted, since they are
        no use as album art.
      operationId: discogsSearch
      parameters:
        - name: q
          in: query
          required: true
          description: Free text, such as an artist and album name.
          schema: { type: string }
        - name: limit
          in: query
          description: |
            How many candidates to fetch covers for. Capped by the server to
            protect the rate limit; the response is authoritative.
          schema: { type: integer, minimum: 1, maximum: 50 }
      responses:
        '200':
          description: Candidate covers.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscogsSearchResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/LookupUnavailable' }
  /discogs/masters/{id}:
    get:
      tags: [discogs]
      summary: Every image on a Discogs master
      description: |
        The front cover is what a search offers, but a master often holds ten
        or more pictures — back cover, inner sleeve, the disc itself. This
        returns all of them, primary first. Served from cache when the master
        was fetched recently, in which case it costs no request.
      operationId: discogsMaster
      parameters:
        - name: id
          in: path
          required: true
          description: A Discogs master id, as returned by the search.
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: The master and its images.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscogsMaster' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/LookupUnavailable' }
...
  /artwork/summary:
    get:
      tags: [artwork]
      summary: Summarise artwork across a selection
      description: |
        Groups identical covers together, which is the only way the duplication
        becomes visible: on a real library the great majority of embedded
        artwork is one image repeated once per track.
      operationId: artworkSummary
      parameters:
        - $ref: '#/components/parameters/Query'
      responses:
        '200':
          description: The summary.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtworkReport' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Required whenever the server is bound to anything other than loopback.
        The token is generated on first run and printed once.
...
  parameters:
    TrackId:
      name: id
      in: path
      required: true
      description: A stable identifier derived from the file's path.
      schema: { type: string }
...